Refactoring the general functionality into a class, separating concerns into different files, creating a main file to call and print a map.
