+++ main.py
@@ -1,0 +1,6 @@
+from map import Map
+
+myMap = Map(10, 10, 0, 0, 9, 9)
+myMap.generateMap()
+
+print(myMap.map)
